% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{analyze_fixedStrategy}
\alias{analyze_fixedStrategy}
\title{analyze_fixedStrategy}
\usage{
analyze_fixedStrategy(
  specifiedAssignmentMechanism = NULL,
  hypotheticalAssignmentMechanism = NULL
)
}
\arguments{
\item{dfm}{'document-feature matrix'. A list ...}
}
\value{
A list consiting of \itemize{
  \item Items.
}
}
\description{
Implements ...
}
\section{References}{

\itemize{
\item Kosuke Imai, Rohit, Connor
}
}

\examples{
#set seed
set.seed(1)

#Geneate data
x <- rnorm(100)

}
