% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeQ_lda.R
\name{computeQ_lda}
\alias{computeQ_lda}
\title{computeQ_lda}
\usage{
computeQ_lda(
  hypotheticalTopicProportion = NULL,
  n_fold = 3,
  Yobs,
  topicProportions,
  documents_list,
  wordTopicDistributions,
  se_ub = sd(Yobs)/10,
  split1_indices = NULL,
  split2_indices = NULL,
  computeThetaSEs = T,
  findMax = T,
  nboot = 10,
  trim_q = 1,
  maxWt = 1e+10,
  maxWt_hajek = NULL,
  alphaLevel = 0.05,
  openBrowser = F
)
}
\arguments{
\item{dfm}{'document-feature matrix'. A list ...}
}
\value{
A list consiting of \itemize{
  \item Items.
}
}
\description{
Implements ...
}
\section{References}{

\itemize{
\item Kosuke Imai, Rohit, Connor
}
}

\examples{
#set seed
set.seed(1)

#Geneate data
x <- rnorm(100)

}
