% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeQse_conjoint.R
\name{computeQse_conjoint}
\alias{computeQse_conjoint}
\title{computeQse_conjoint}
\usage{
computeQse_conjoint(theta, FactorMatrix, log_pr_w = NULL)
}
\arguments{
\item{dfm}{'document-feature matrix'. A list ...}
}
\value{
A list consiting of \itemize{
  \item Items.
}
}
\description{
Implements ...
}
\section{References}{

\itemize{
\item Kosuke, Rohit, Connor.  Working Paper.
}
}

\examples{
#set seed
set.seed(1)

#Geneate data
x <- rnorm(100)

}
