% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeQse_lda.R
\name{computeQse_lda}
\alias{computeQse_lda}
\title{computeQse_lda}
\usage{
computeQse_lda(
  THETA__,
  INDICES_,
  DOC_INDICES_U,
  D_INDICES_U,
  PI_MAT_INPUT,
  MARGINAL_BOUNDS,
  DOC_LIST,
  MODAL_DOC_LEN,
  TERMS_MAT_INPUT,
  LOG_TREATCOMBS,
  YOBS,
  returnLog = T,
  LOG_PR_W = NULL
)
}
\arguments{
\item{dfm}{'document-feature matrix'. A list ...}
}
\value{
A list consiting of \itemize{
  \item Items.
}
}
\description{
Implements ...
}
\section{References}{

\itemize{
\item Kosuke, Rohit, Connor.  Working Paper.
}
}

\examples{
#set seed
set.seed(1)

#Geneate data
x <- rnorm(100)

}
