% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{find_optimalStrategy}
\alias{find_optimalStrategy}
\title{find_optimalStrategy}
\usage{
find_optimalStrategy(specifiedAssignmentMechanism = NULL, cubeConstraint = F)
}
\arguments{
\item{dfm}{'document-feature matrix'. A list ...}
}
\value{
A list consiting of \itemize{
  \item Items.
}
}
\description{
Implements ...
}
\section{References}{

\itemize{
\item Kosuke Imai, Rohit, Connor
}
}

\examples{
#set seed
set.seed(1)

#Geneate data
x <- rnorm(100)

}
