% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{optimizecausalities}
\alias{optimizecausalities}
\title{optimizecausalities}
\usage{
optimizecausalities(Yobs, W, optimize = T, theta = NULL)
}
\arguments{
\item{dfm}{'document-feature matrix'. A list ...}
}
\value{
A list consiting of \itemize{
  \item Items.
}
}
\description{
Implements ...
}
\section{References}{

\itemize{
\item Kosuke Imai, Rohit, Connor
}
}

\examples{
#set seed
set.seed(1)

#Geneate data
x <- rnorm(100)

}
