% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizeQ_lda.R
\name{optimizeQ_lda}
\alias{optimizeQ_lda}
\title{computeQ_lda}
\usage{
optimizeQ_lda(
  INDICES_SPLIT1 = NULL,
  INDICES_SPLIT2 = NULL,
  DTM_MAT,
  n_fold = 3,
  YOBS,
  PI_MAT,
  DOC_LIST,
  TERMS_MAT,
  SE_UB = sd(YOBS)/10,
  nboot = 10,
  trim_q = 1,
  maxWt = 1e+10,
  maxWt_hajek = NULL,
  computeSEs = T,
  doMax = T,
  alphaLevel = 0.05,
  openBrowser = F
)
}
\arguments{
\item{dfm}{'document-feature matrix'. A list ...}
}
\value{
A list consiting of \itemize{
  \item Items.
}
}
\description{
Implements ...
}
\section{References}{

\itemize{
\item Kosuke Imai, Rohit, Connor
}
}

\examples{
#set seed
set.seed(1)

#Geneate data
x <- rnorm(100)

}
