% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizeQ_conjoint.R
\name{computeQ_conjoint}
\alias{computeQ_conjoint}
\title{computeQ_conjoint}
\usage{
computeQ_conjoint(
  FactorsMat,
  Yobs,
  assignmentProbList,
  hypotheticalProbList = NULL,
  se_ub,
  INDICES_SPLIT1 = NULL,
  INDICES_SPLIT2 = NULL,
  computeSEs = F,
  openBrowser = F,
  hajek = T,
  doMax = T,
  quiet = T
)
}
\arguments{
\item{dfm}{'document-feature matrix'. A list ...}
}
\value{
A list consiting of \itemize{
  \item Items.
}
}
\description{
Implements ...
}
\section{References}{

\itemize{
\item Kosuke Imai, Rohit, Connor
}
}

\examples{
#set seed
set.seed(1)

#Geneate data
x <- rnorm(100)

}
