% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeQ_lda.R
\name{computeQ_lda}
\alias{computeQ_lda}
\title{computeQ_lda}
\usage{
computeQ_lda(
  theta = NULL,
  term_mat,
  Yobs,
  doc_words,
  dtm = NULL,
  pi_mat = NULL,
  alpha_mat = NULL,
  log_pr_w = NULL,
  computeSE = F,
  trim_q = 1,
  quiet = T,
  iters = 100,
  smoothWts = F,
  TreatFxn = NULL,
  maxWt = 1e+10,
  maxWt_hajek = NULL,
  term_mat_TRUE = NULL,
  doc_indices_u = NULL,
  d_indices_u = NULL,
  diagnostics = F
)
}
\arguments{
\item{dfm}{'document-feature matrix'. A list ...}
}
\value{
A list consiting of \itemize{
  \item Items.
}
}
\description{
Implements ...
}
\section{References}{

\itemize{
\item Kosuke, Rohit, Connor.  Working Paper.
}
}

\examples{
#set seed
set.seed(1)

#Geneate data
x <- rnorm(100)

}
