% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizeQ_conjoint.R
\name{optimizeQ_conjoint}
\alias{optimizeQ_conjoint}
\title{computeQ_conjoint}
\usage{
optimizeQ_conjoint(FactorMatrix, SE_bound)
}
\arguments{
\item{dfm}{'document-feature matrix'. A list ...}
}
\value{
A list consiting of \itemize{
  \item Items.
}
}
\description{
Implements ...
}
\section{References}{

\itemize{
\item Kosuke Imai, Rohit, Connor
}
}

\examples{
#set seed
set.seed(1)

#Geneate data
x <- rnorm(100)

}
